refactor(cli): move scheduler sub-app to app/cli/scheduler.py
Moves scheduler_start_cmd and scheduler_status_cmd. The status command
keeps delegating to schedule_list_cmd, which is now imported from
cli.schedule directly. main.py drops the temporary schedule_list_cmd
re-import that Task 9 added.
